Add feature branch to dataflows plugin
Despite confirmation from Catalyst that the new dataflow should without any modified code, it does not work in prod where it has the latest stable branch. This is a test to see if it works using the experimental branch (to suit an older version of the new flow)  that we're currently running in dev and that works, works. If this doesn't work, then something else is happening in prod and I don't know how to proceed.